Add support for overriding jira issues to comment on. (#15427)
(cherry picked from commit 91f875b16be7fba71bb1f80ee2563724f2aca808)
